q2/requirements.txt
# FastAPI and ASGI server
fastmcp==2.9.2
uvicorn==0.29.0

# Time handling
pytz==2024.1
python-dateutil==2.9.0.post0

# Data processing
pandas==2.2.2

# Gemini API (Google Generative AI SDK)
google-generativeai==0.5.3

# Environment variable handling
python-dotenv==1.0.1

# Testing
pytest==8.2.1
httpx==0.27.0
